Refactor duplicate methods
Make the game result ("1-0", "1/2-1/2", etc.) a method of model.Game
so it can be used in Xboard communication and PGN writing.
